requirements.txt
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
redis>=4.0.0
